Update Testfile name.
* Json String is read from the file instead of variable.
